ajout des metier

+++ artisanJpaBack/src/main/java/jpa/dao/DaoService.java
@@ -2,6 +2,6 @@ package jpa.dao;
 
 import jpa.model.Service;
 
-public interface DaoService extends DaoGeneric<Service, Integer>{
+public interface DaoService extends DaoGeneric<Service, Long>{
 
 }
